refactor transaction creation methods into own service
